+++ .eslintrc.js
@@ -1,7 +1,7 @@
 module.exports = {
   env: {
     browser: true,
-    es2022: true,
+    // es2022: true,
     commonjs: true,
     node: true,
   },