+++ example/static/static-a/vue.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+    publicPath:"http://localhost:9000",
+    devServer: {
+      "port":9000,
+      "headers": {
+        "Access-Control-Allow-Origin": "*"
+      }
+    }
+}