+++ packages/local-mock-middleware/src/index.ts
@@ -1,16 +1,3 @@
-import { generateTemplate } from '@/core/constants'
-import expressMiddleware from '@/core/express-middleware'
-import koaMiddleware from '@/core/koa-middleware'
+import createLockMock from './core/local-mock'
 
-export { generateTemplate, koaMiddleware, expressMiddleware }
-
-export const name = 'local-mock-middleware'
-
-const middleware = {
-  name,
-  generateTemplate,
-  expressMiddleware,
-  koaMiddleware,
-}
-
-export default middleware
+export default createLockMock