+++ packages/local-mock-easy/CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/vigory/local-mock-core/compare/local-mock-easy@1.0.7-beta.2...local-mock-easy@1.1.0) (2022-12-16)
+
+### Bug Fixes
+
+- rollup.config.ts ([5039e57](https://github.com/vigory/local-mock-core/commit/5039e576702f609e9995e0f7a9c6183107dd9bd0))
+
+### Features
+
+- 可滑动 ([b0bf645](https://github.com/vigory/local-mock-core/commit/b0bf645320a44c3a48312008d1843526325438b3))
+
+## [1.0.7](https://github.com/vigory/local-mock-core/compare/local-mock-easy@1.0.7-beta.2...local-mock-easy@1.0.7) (2022-11-25)
+
+**Note:** Version bump only for package local-mock-easy
+
+## [1.0.7-beta.2](https://github.com/vigory/local-mock-core/compare/local-mock-easy@1.0.7-beta.0...local-mock-easy@1.0.7-beta.2) (2022-11-24)
+
+**Note:** Version bump only for package local-mock-easy
+
+## [1.0.7-beta.1](https://github.com/vigory/local-mock-core/compare/local-mock-easy@1.0.7-beta.0...local-mock-easy@1.0.7-beta.1) (2022-11-24)
+
+**Note:** Version bump only for package local-mock-easy
+
 ## [1.0.7-beta.0](https://github.com/vigory/local-mock-core/compare/local-mock-easy@1.0.6...local-mock-easy@1.0.7-beta.0) (2022-11-22)
 
 ### Bug Fixes
@@ -18,13 +40,13 @@ All notable changes to this project will be documented in this file. See [Conven
 
 **Note:** Version bump only for package local-mock-easy
 
-## [1.0.4](https://github.com/redstar08/local-mock-core/compare/local-mock-easy@1.0.3...local-mock-easy@1.0.4) (2022-11-19)
+## [1.0.4](https://github.com/vigory/local-mock-core/compare/local-mock-easy@1.0.3...local-mock-easy@1.0.4) (2022-11-19)
 
 ### Features
 
-- update docs ([efeb7b1](https://github.com/redstar08/local-mock-core/commit/efeb7b1d0a7fae9bf7d6aed545dec825728291a0))
+- update docs ([efeb7b1](https://github.com/vigory/local-mock-core/commit/efeb7b1d0a7fae9bf7d6aed545dec825728291a0))
 
-## [1.0.3](https://github.com/redstar08/local-mock-core/compare/local-mock-easy@1.0.2...local-mock-easy@1.0.3) (2022-11-14)
+## [1.0.3](https://github.com/vigory/local-mock-core/compare/local-mock-easy@1.0.2...local-mock-easy@1.0.3) (2022-11-14)
 
 **Note:** Version bump only for package local-mock-easy
 
@@ -32,4 +54,4 @@ All notable changes to this project will be documented in this file. See [Conven
 
 ### Features
 
-- update local-mock-easy ([7994f66](https://github.com/redstar08/local-mock-core/commit/7994f663e425a0608a502f676e2c3f111186df35))
+- update local-mock-easy ([7994f66](https://github.com/vigory/local-mock-core/commit/7994f663e425a0608a502f676e2c3f111186df35))