+++ packages/local-mock-middleware/CHANGELOG.md
@@ -2,6 +2,40 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/vigory/local-mock-core/compare/local-mock-middleware@1.0.8...local-mock-middleware@1.1.0) (2022-12-16)
+
+### Bug Fixes
+
+- rollup.config.ts ([5039e57](https://github.com/vigory/local-mock-core/commit/5039e576702f609e9995e0f7a9c6183107dd9bd0))
+
+## [1.0.8](https://github.com/vigory/local-mock-core/compare/local-mock-middleware@1.0.7-beta.4...local-mock-middleware@1.0.8) (2022-11-25)
+
+**Note:** Version bump only for package local-mock-middleware
+
+## [1.0.7-beta.4](https://github.com/vigory/local-mock-core/compare/local-mock-middleware@1.0.7-beta.2...local-mock-middleware@1.0.7-beta.4) (2022-11-25)
+
+### Bug Fixes
+
+- 修复 koa 无法路径判断问题 ([206bac3](https://github.com/vigory/local-mock-core/commit/206bac358e83be8c3d9cc06fee94cdd3db866692))
+
+## [1.0.7-beta.3](https://github.com/vigory/local-mock-core/compare/local-mock-middleware@1.0.7-beta.2...local-mock-middleware@1.0.7-beta.3) (2022-11-24)
+
+### Bug Fixes
+
+- 修复 koa 无法路径判断问题 ([206bac3](https://github.com/vigory/local-mock-core/commit/206bac358e83be8c3d9cc06fee94cdd3db866692))
+
+## [1.0.7-beta.2](https://github.com/vigory/local-mock-core/compare/local-mock-middleware@1.0.7-beta.0...local-mock-middleware@1.0.7-beta.2) (2022-11-24)
+
+### Features
+
+- update ([4484b60](https://github.com/vigory/local-mock-core/commit/4484b60a57704bfd280232311ed42c68fb99673f))
+
+## [1.0.7-beta.1](https://github.com/vigory/local-mock-core/compare/local-mock-middleware@1.0.7-beta.0...local-mock-middleware@1.0.7-beta.1) (2022-11-24)
+
+### Features
+
+- update ([4484b60](https://github.com/vigory/local-mock-core/commit/4484b60a57704bfd280232311ed42c68fb99673f))
+
 ## [1.0.7-beta.0](https://github.com/vigory/local-mock-core/compare/local-mock-middleware@1.0.6...local-mock-middleware@1.0.7-beta.0) (2022-11-22)
 
 **Note:** Version bump only for package local-mock-middleware
@@ -16,20 +50,20 @@ All notable changes to this project will be documented in this file. See [Conven
 
 **Note:** Version bump only for package local-mock-middleware
 
-## [1.0.4](https://github.com/redstar08/local-mock-core/compare/local-mock-middleware@1.0.3...local-mock-middleware@1.0.4) (2022-11-19)
+## [1.0.4](https://github.com/vigory/local-mock-core/compare/local-mock-middleware@1.0.3...local-mock-middleware@1.0.4) (2022-11-19)
 
 ### Features
 
-- add isOpen option ([2aa5b96](https://github.com/redstar08/local-mock-core/commit/2aa5b96f4dbf85bbe5bd9cf5413b89ac1455d891))
-- export generateTemplate to custom middleware ([e999796](https://github.com/redstar08/local-mock-core/commit/e9997965391b8740ff9dc7908d44457e0d739cc5))
-- update docs ([efeb7b1](https://github.com/redstar08/local-mock-core/commit/efeb7b1d0a7fae9bf7d6aed545dec825728291a0))
+- add isOpen option ([2aa5b96](https://github.com/vigory/local-mock-core/commit/2aa5b96f4dbf85bbe5bd9cf5413b89ac1455d891))
+- export generateTemplate to custom middleware ([e999796](https://github.com/vigory/local-mock-core/commit/e9997965391b8740ff9dc7908d44457e0d739cc5))
+- update docs ([efeb7b1](https://github.com/vigory/local-mock-core/commit/efeb7b1d0a7fae9bf7d6aed545dec825728291a0))
 
-## [1.0.3](https://github.com/redstar08/local-mock-core/compare/local-mock-middleware@1.0.2...local-mock-middleware@1.0.3) (2022-11-14)
+## [1.0.3](https://github.com/vigory/local-mock-core/compare/local-mock-middleware@1.0.2...local-mock-middleware@1.0.3) (2022-11-14)
 
 **Note:** Version bump only for package local-mock-middleware
 
-## [1.0.2](https://github.com/redstar08/local-mock-core/compare/local-mock-middleware@1.0.1...local-mock-middleware@1.0.2) (2022-11-14)
+## [1.0.2](https://github.com/vigory/local-mock-core/compare/local-mock-middleware@1.0.1...local-mock-middleware@1.0.2) (2022-11-14)
 
 ### Features
 
-- update local-mock-easy ([7994f66](https://github.com/redstar08/local-mock-core/commit/7994f663e425a0608a502f676e2c3f111186df35))
+- update local-mock-easy ([7994f66](https://github.com/vigory/local-mock-core/commit/7994f663e425a0608a502f676e2c3f111186df35))